feat: 메인페이지 카테고리 카드에서 카테고리 클릭시, 카테고리 검색 필터가 적용되도록 구현
# Conflicts:
#	src/app/_containers/mainpage-content.tsx
